Move to just the AWST.
