+++ src/Network/AWS/Flow/Types.hs
@@ -1,13 +1,22 @@
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE ConstraintKinds            #-}
+{-# LANGUAGE FlexibleInstances          #-}
+{-# LANGUAGE MultiParamTypeClasses      #-}
+{-# LANGUAGE UndecidableInstances       #-}
+{-# LANGUAGE TypeFamilies               #-}
 
 module Network.AWS.Flow.Types where
 
+import Control.Lens
+import Control.Monad.Base
 import Control.Monad.Catch
+import Control.Monad.Except
 import Control.Monad.Logger
 import Control.Monad.Reader
+import Control.Monad.Trans.Control
 import Control.Monad.Trans.Resource
 import Control.Monad.Trans.AWS
+import Data.Aeson
 import Data.ByteString.Lazy
 import Data.Conduit.Lazy
 import Data.Text
@@ -45,7 +54,7 @@ data FlowEnv = FlowEnv
   }
 
 newtype FlowT m a = FlowT
-  { unFlowT :: ReaderT FlowEnv m a
+  { unFlowT :: AWST' FlowEnv m a
   } deriving ( Functor
              , Applicative
              , Monad
@@ -55,11 +64,50 @@ newtype FlowT m a = FlowT
              )
 
 type MonadFlow m =
-  ( MonadThrow m
-  , MonadCatch m
+  ( MonadCatch m
+  , MonadThrow m
   , MonadResource m
   , MonadReader FlowEnv m
   )
+
+instance MonadThrow m => MonadThrow (FlowT m) where
+    throwM = lift . throwM
+
+instance MonadCatch m => MonadCatch (FlowT m) where
+    catch (FlowT m) f = FlowT (catch m (unFlowT . f))
+
+instance MonadBase b m => MonadBase b (FlowT m) where
+    liftBase = liftBaseDefault
+
+instance MonadTransControl FlowT where
+    type StT FlowT a = StT (ReaderT FlowEnv) a
+
+    liftWith = defaultLiftWith FlowT unFlowT
+    restoreT = defaultRestoreT FlowT
+
+instance MonadBaseControl b m => MonadBaseControl b (FlowT m) where
+    type StM (FlowT m) a = ComposeSt FlowT m a
+
+    liftBaseWith = defaultLiftBaseWith
+    restoreM     = defaultRestoreM
+
+instance MonadResource m => MonadResource (FlowT m) where
+    liftResourceT = lift . liftResourceT
+
+instance MonadError e m => MonadError e (FlowT m) where
+    throwError     = lift . throwError
+    catchError m f = FlowT (catchError (unFlowT m) (unFlowT . f))
+
+instance Monad m => MonadReader FlowEnv (FlowT m) where
+    ask     = FlowT ask
+    local f = FlowT . local f . unFlowT
+    reader  = FlowT . reader
+
+instance HasEnv FlowEnv where
+  environment = lens feEnv (\s a -> s { feEnv = a })
+
+runFlowT :: FlowEnv -> FlowT m a -> m a
+runFlowT e (FlowT m) = runAWST e m
 
 data DecideEnv = DecideEnv
   { deLogger    :: Log
@@ -80,9 +128,46 @@ newtype DecideT m a = DecideT
 
 type MonadDecide m =
   ( MonadCatch m
+  , MonadThrow m
   , MonadResource m
   , MonadReader DecideEnv m
   )
+
+instance MonadThrow m => MonadThrow (DecideT m) where
+    throwM = lift . throwM
+
+instance MonadCatch m => MonadCatch (DecideT m) where
+    catch (DecideT m) f = DecideT (catch m (unDecideT . f))
+
+instance MonadBase b m => MonadBase b (DecideT m) where
+    liftBase = liftBaseDefault
+
+instance MonadTransControl DecideT where
+    type StT DecideT a = StT (ReaderT DecideEnv) a
+
+    liftWith = defaultLiftWith DecideT unDecideT
+    restoreT = defaultRestoreT DecideT
+
+instance MonadBaseControl b m => MonadBaseControl b (DecideT m) where
+    type StM (DecideT m) a = ComposeSt DecideT m a
+
+    liftBaseWith = defaultLiftBaseWith
+    restoreM     = defaultRestoreM
+
+instance MonadResource m => MonadResource (DecideT m) where
+    liftResourceT = lift . liftResourceT
+
+instance MonadError e m => MonadError e (DecideT m) where
+    throwError     = lift . throwError
+    catchError m f = DecideT (catchError (unDecideT m) (unDecideT . f))
+
+instance Monad m => MonadReader DecideEnv (DecideT m) where
+    ask     = DecideT ask
+    local f = DecideT . local f . unDecideT
+    reader  = DecideT . reader
+
+runDecideT :: DecideEnv -> DecideT m a -> m a
+runDecideT e (DecideT m) = runReaderT m e
 
 data Task = Task
   { tskName    :: Name
@@ -91,14 +176,36 @@ data Task = Task
   , tskTimeout :: Timeout
   } deriving ( Eq, Read, Show )
 
+instance FromJSON Task where
+  parseJSON (Object v) =
+    Task             <$>
+      v .: "name"    <*>
+      v .: "version" <*>
+      v .: "queue"   <*>
+      v .: "timeout"
+  parseJSON _ = mzero
+
 data Timer = Timer
   { tmrName    :: Name
   , tmrTimeout :: Timeout
   } deriving ( Eq, Read, Show )
 
+instance FromJSON Timer where
+  parseJSON (Object v) =
+    Timer            <$>
+      v .: "name"    <*>
+      v .: "timeout"
+  parseJSON _ = mzero
+
 data Start = Start
   { strtTask :: Task
   } deriving ( Eq, Read, Show )
+
+instance FromJSON Start where
+  parseJSON (Object v) =
+    Start         <$>
+      v .: "flow"
+  parseJSON _ = mzero
 
 data Spec
   = Work
@@ -108,13 +215,40 @@ data Spec
   { slpTimer :: Timer
   } deriving ( Eq, Read, Show )
 
+instance FromJSON Spec where
+  parseJSON (Object v) =
+    msum
+      [ Work           <$>
+          v .: "work"
+      , Sleep          <$>
+          v .: "sleep"
+      ]
+  parseJSON _ =
+    mzero
+
 data End
   = Stop
   | Continue
   deriving ( Eq, Read, Show )
+
+instance FromJSON End where
+  parseJSON (String v)
+    | v == "stop"     = return Stop
+    | v == "continue" = return Continue
+    | otherwise = mzero
+  parseJSON _ = mzero
 
 data Plan = Plan
   { plnStart :: Start
   , plnSpecs :: [Spec]
   , plnEnd   :: End
   } deriving ( Eq, Read, Show )
+
+instance FromJSON Plan where
+  parseJSON (Object v) =
+    Plan           <$>
+      v .: "start" <*>
+      v .: "specs" <*>
+      v .: "end"
+  parseJSON _ = mzero
+
