+++ wolf.cabal
@@ -21,8 +21,7 @@ source-repository head
 
 library
   exposed-modules:     Network.AWS.Flow
-                     , Network.AWS.Flow.Env
-  other-modules:       Network.AWS.Flow.Internal
+  other-modules:       Network.AWS.Flow.Env
                      , Network.AWS.Flow.S3
                      , Network.AWS.Flow.SWF
                      , Network.AWS.Flow.Types
